git-svn-id: http://svn.codehaus.org/groovy/branches/GROOVY_1_7_X@20584 a5544e8c-8a19-0410-ba12-f9af4593a198

+++ src/main/org/codehaus/groovy/runtime/dgmimpl/arrays/BooleanArrayGetAtMetaMethod.java
@@ -23,27 +23,27 @@ import org.codehaus.groovy.runtime.callsite.CallSite;
 import org.codehaus.groovy.runtime.callsite.PojoMetaMethodSite;
 
 public class BooleanArrayGetAtMetaMethod extends ArrayGetAtMetaMethod {
-   private static final CachedClass ARR_CLASS = ReflectionCache.getCachedClass(boolean[].class);
+    private static final CachedClass ARR_CLASS = ReflectionCache.getCachedClass(boolean[].class);
 
-   public Class getReturnType() {
-       return Boolean.class;
-   }
+    public Class getReturnType() {
+        return Boolean.class;
+    }
 
-   public final CachedClass getDeclaringClass() {
-       return ARR_CLASS;
-   }
+    public final CachedClass getDeclaringClass() {
+        return ARR_CLASS;
+    }
 
-   public Object invoke(Object object, Object[] args) {
-       final boolean[] objects = (boolean[]) object;
-       return objects[normaliseIndex(((Integer) args[0]).intValue(), objects.length)];
-   }
+    public Object invoke(Object object, Object[] args) {
+        final boolean[] objects = (boolean[]) object;
+        return objects[normaliseIndex((Integer) args[0], objects.length)];
+    }
 
-   public CallSite createPojoCallSite(CallSite site, MetaClassImpl metaClass, MetaMethod metaMethod, Class[] params, Object receiver, Object[] args) {
-       if (!(args [0] instanceof Integer))
-         return PojoMetaMethodSite.createNonAwareCallSite(site, metaClass, metaMethod, params, args);
-       else
-           return new MyPojoMetaMethodSite(site, metaClass, metaMethod, params);
-   }
+    public CallSite createPojoCallSite(CallSite site, MetaClassImpl metaClass, MetaMethod metaMethod, Class[] params, Object receiver, Object[] args) {
+        if (!(args[0] instanceof Integer))
+            return PojoMetaMethodSite.createNonAwareCallSite(site, metaClass, metaMethod, params, args);
+        else
+            return new MyPojoMetaMethodSite(site, metaClass, metaMethod, params);
+    }
 
     private static class MyPojoMetaMethodSite extends PojoMetaMethodSite {
         public MyPojoMetaMethodSite(CallSite site, MetaClassImpl metaClass, MetaMethod metaMethod, Class[] params) {
@@ -54,10 +54,9 @@ public class BooleanArrayGetAtMetaMethod extends ArrayGetAtMetaMethod {
             if ((receiver instanceof boolean[] && arg instanceof Integer)
                     && checkPojoMetaClass()) {
                 final boolean[] objects = (boolean[]) receiver;
-                return objects[normaliseIndex(((Integer) arg).intValue(), objects.length)];
-            }
-            else
-              return super.call(receiver,arg);
+                return objects[normaliseIndex((Integer) arg, objects.length)];
+            } else
+                return super.call(receiver, arg);
         }
     }
 }
